zomato-price-prediction/requirements.txt
Flask
pandas
scikit-learn
mysql-connector-python
